backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
sqlalchemy==2.0.25
llama-index==0.10.15
llama-index-llms-gemini==0.1.3
llama-index-embeddings-gemini==0.1.3
llama-index-embeddings-huggingface==0.6.1
llama-index-vector-stores-chroma==0.1.4
llama-index-llms-anthropic==0.10.11
chromadb==0.4.24
python-dotenv==1.0.1
pandas==2.2.0
google-genai
